backend/requirements.txt
django>=5.0
djangorestframework>=3.14
django-cors-headers>=4.3
django-environ>=0.11
psycopg2-binary>=2.9.0
dj-database-url>=2.0.0
gunicorn>=21.2.0
whitenoise>=6.5.0
google-genai>=1.0.0
langgraph>=0.2.0
langchain-google-genai>=2.0.0
pytest-django>=4.7
pytest>=8.0
pytest-asyncio>=0.21
# Epic 3: ChromaDB Vector Store
chromadb>=0.4.0
sentence-transformers>=2.2.0
# Epic 5: PDF Export
weasyprint>=60.0
markdown>=3.5
# OSINT feature
dnspython>=2.4
python-whois>=0.9
python-docx>=1.1
httpx>=0.27
Pillow>=10.0
